Subprocessing functions moved to a utils file + nodes also using WITH + UNWIND statement
